Move AstraMisc.cpp into AstraMisc.h
Not necessary at all to keep these seperate, all function code can go in AstraMisc.h
